Merge In Upstream Changes

+++ ci/build_all_nix.sh
@@ -2,6 +2,7 @@
 
 #TODO(kizill): split this into subscripts to make it prettier
 
+eval "$(pyenv init --path)"
 eval "$(pyenv init -)"
 eval "$(pyenv virtualenv-init -)"
 
@@ -23,6 +24,7 @@ function python_version {
         Python*3.7*) echo 3.7 ;;
         Python*3.8*) echo 3.8 ;;
         Python*3.9*) echo 3.9 ;;
+        Python*3.10*) echo 3.10 ;;
         *) echo "Cannot determine python version" ; exit 1 ;;
     esac
 }
@@ -30,7 +32,7 @@ function python_version {
 function os_sdk {
     python_version=`python_version python`
     case `uname -s` in
-        Linux) echo "-DOS_SDK=ubuntu-10 -DUSE_SYSTEM_PYTHON=$python_version" ;;
+        Linux) echo "-DOS_SDK=ubuntu-12 -DUSE_SYSTEM_PYTHON=$python_version" ;;
         *) echo "-DOS_SDK=local" ;;
     esac
 }
@@ -58,7 +60,8 @@ cp -r tests catboost
 python ../../ya make -r $lnx_common_flags $(os_sdk) -T src
 
 mkdir -p catboost/inst/libs
-cp $(readlink src/libcatboostr.so) catboost/inst/libs
+[ -s "src/libcatboostr.so" ] && cp $(readlink src/libcatboostr.so) catboost/inst/libs
+[ -s "src/libcatboostr.dylib" ] && cp $(readlink src/libcatboostr.dylib) catboost/inst/libs
 
 tar -cvzf catboost-R-$(uname).tgz catboost
 
@@ -92,6 +95,11 @@ python mk_wheel.py $lnx_common_flags $(os_sdk) -DPYTHON_CONFIG=$(pyenv prefix)/b
 PY39=3.9.0
 pyenv install -s $PY39
 pyenv shell $PY39
+python mk_wheel.py $lnx_common_flags $(os_sdk) -DPYTHON_CONFIG=$(pyenv prefix)/bin/python3-config
+
+PY310=3.10.0
+pyenv install -s $PY310
+pyenv shell $PY310
 python mk_wheel.py $lnx_common_flags $(os_sdk) -DPYTHON_CONFIG=$(pyenv prefix)/bin/python3-config
